Ooopps, grave error related to trail_lim
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/cryptominisat/branches/gaussian_differentmatrix@686 6c43cb19-1bce-4953-9960-ac6414807ffa

+++ Solver/Solver.C
@@ -394,7 +394,7 @@ void Solver::cancelUntil(int level)
         for (Gaussian **gauss = &gauss_matrixes[0], **end= gauss + gauss_matrixes.size(); gauss != end; gauss++)
             (*gauss)->canceling(trail_lim[level]);
         
-        for (int c = trail.size()-1; c >= trail_lim[level]; c--) {
+        for (int c = trail.size()-1; c >= (int)trail_lim[level]; c--) {
             Var     x  = trail[c].var();
             #ifdef VERBOSE_DEBUG
             cout << "Canceling var " << x+1 << " sublevel: " << c << endl;
